Create a VM class
