+++ gns3/vm.py
@@ -1,0 +1,157 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright (C) 2014 GNS3 Technologies Inc.
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+"""
+Base class for VM classes.
+"""
+
+from .qt import QtCore
+from .node import Node
+
+import logging
+log = logging.getLogger(__name__)
+
+
+class VM(Node):
+
+    def __init__(self, module, server, project):
+
+        super().__init__(module, server, project)
+
+        self._vm_id = None
+
+    def vm_id(self):
+        """
+        Return the ID of this device
+
+        :returns: identifier (string)
+        """
+
+        return self._vm_id
+
+    def delete(self):
+        """
+        Deletes this VPCS instance.
+        """
+
+        log.debug("VPCS device {} is being deleted".format(self.name()))
+        # first delete all the links attached to this node
+        self.delete_links_signal.emit()
+        if self._vm_id:
+            self.httpDelete("/vpcs/vms/{vm_id}".format(project_id=self._project.id(), vm_id=self._vm_id), self._deleteCallback)
+        else:
+            self.deleted_signal.emit()
+            self._module.removeNode(self)
+
+    def _deleteCallback(self, result, error=False, **kwargs):
+        """
+        Callback for delete.
+
+        :param result: server response (dict)
+        :param error: indicates an error (boolean)
+        """
+
+        if error:
+            log.error("error while deleting {}: {}".format(self.name(), result["message"]))
+            self.server_error_signal.emit(self.id(), result["message"])
+        log.info("{} has been deleted".format(self.name()))
+        self.deleted_signal.emit()
+        self._module.removeNode(self)
+
+    def start(self):
+        """
+        Starts this VPCS instance.
+        """
+
+        if self.status() == Node.started:
+            log.debug("{} is already running".format(self.name()))
+            return
+
+        log.debug("{} is starting".format(self.name()))
+        self.httpPost("/{prefix}/vms/{vm_id}/start".format(prefix=self.URL_PREFIX, vm_id=self._vm_id), self._startCallback)
+
+    def _startCallback(self, result, error=False, **kwargs):
+        """
+        Callback for start.
+
+        :param result: server response (dict)
+        :param error: indicates an error (boolean)
+        """
+
+        if error:
+            log.error("error while starting {}: {}".format(self.name(), result["message"]))
+            self.server_error_signal.emit(self.id(), result["message"])
+        else:
+            log.info("{} has started".format(self.name()))
+            self.setStatus(Node.started)
+            for port in self._ports:
+                # set ports as started
+                port.setStatus(Port.started)
+            self.started_signal.emit()
+
+    def stop(self):
+        """
+        Stops this VPCS instance.
+        """
+
+        if self.status() == Node.stopped:
+            log.debug("{} is already stopped".format(self.name()))
+            return
+
+        log.debug("{} is stopping".format(self.name()))
+        self.httpPost("/{prefix}/vms/{vm_id}/stop".format(prefix=self.URL_PREFIX, vm_id=self._vm_id), self._stopCallback)
+
+    def _stopCallback(self, result, error=False, **kwargs):
+        """
+        Callback for stop.
+
+        :param result: server response (dict)
+        :param error: indicates an error (boolean)
+        """
+
+        if error:
+            log.error("error while stopping {}: {}".format(self.name(), result["message"]))
+            self.server_error_signal.emit(self.id(), result["message"])
+        else:
+            log.info("{} has stopped".format(self.name()))
+            self.setStatus(Node.stopped)
+            for port in self._ports:
+                # set ports as stopped
+                port.setStatus(Port.stopped)
+            self.stopped_signal.emit()
+
+    def reload(self):
+        """
+        Reloads this VPCS instance.
+        """
+
+        log.debug("{} is being reloaded".format(self.name()))
+        self.httpPost("/{prefix}/vms/{vm_id}/reload".format(prefix=self.URL_PREFIX, vm_id=self._vm_id), self._reloadCallback)
+
+    def _reloadCallback(self, result, error=False, **kwargs):
+        """
+        Callback for reload.
+
+        :param result: server response (dict)
+        :param error: indicates an error (boolean)
+        """
+
+        if error:
+            log.error("error while suspending {}: {}".format(self.name(), result["message"]))
+            self.server_error_signal.emit(self.id(), result["message"])
+        else:
+            log.info("{} has reloaded".format(self.name()))
